+++ commit/trk2dictionary/trk2dictionary.pyx
@@ -18,7 +18,7 @@ from libcpp cimport bool
 # Interface to actual C code
 cdef extern from "trk2dictionary_c.cpp":
     int trk2dictionary(
-        char* filename_tractogram, int data_offset, int Nx, int Ny, int Nz, float Px, float Py, float Pz, int n_count, int n_scalars, 
+        char* filename_tractogram, int data_offset, int Nx, int Ny, int Nz, float Px, float Py, float Pz, int n_count, int n_scalars,
         int n_properties, float fiber_shiftX, float fiber_shiftY, float fiber_shiftZ, float min_seg_len, float min_fiber_len,  float max_fiber_len,
         float* ptrPEAKS, int Np, float vf_THR, int ECix, int ECiy, int ECiz,
         float* _ptrMASK, float* ptrTDI, char* path_out, int c, double* ptrPeaksAffine,
@@ -26,10 +26,15 @@ cdef extern from "trk2dictionary_c.cpp":
         float* ptrTractsAffine, unsigned short ndirs, short* prtHashTable
     ) nogil
 
+def _get_header( niiFILE ):
+    return niiFILE.header if nibabel.__version__ >= '2.0.0' else niiFILE.get_header()
+
+def _get_affine( niiFILE ):
+    return niiFILE.affine if nibabel.__version__ >= '2.0.0' else niiFILE.get_affine()
 
 cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filename_mask=None, do_intersect=True,
     fiber_shift=0, min_seg_len=1e-3, min_fiber_len=0.0, max_fiber_len=250.0,
-    vf_THR=0.1, peaks_use_affine=False, flip_peaks=[False,False,False], 
+    vf_THR=0.1, peaks_use_affine=False, flip_peaks=[False,False,False],
     blur_spacing=0.25, blur_core_extent=0.0, blur_gauss_extent=0.0, blur_gauss_min=0.1, blur_apply_to=None,
     filename_trk=None, gen_trk=None, TCK_ref_image=None, ndirs=32761
     ):
@@ -41,12 +46,12 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     ----------
     filename_tractogram : string
         Path to the tractogram (.trk or .tck) containing the streamlines to load.
-        
+
     TCK_ref_image: string
         When loading a .tck tractogram, path to the NIFTI file containing the information about
         the geometry to be used for the tractogram to load. If not specified, it will try to use
         the information from filename_peaks or filename_mask.
-    
+
     path_out : string
         Path to the folder for storing the sparse data structure. If not specified (default),
         a folder name "COMMIT" will be created in the same folder of the tractogram.
@@ -57,11 +62,11 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
         the mask is created from all voxels intersected by the tracts.
 
     do_intersect : boolean
-        If True then fiber segments that intersect voxel boundaries are splitted (default).
+        If True then streamline segments that intersect voxel boundaries are splitted (default).
         If False then the centroid of the segment is used as its voxel position.
 
     fiber_shift : float or list of three float
-        If necessary, apply a translation to fiber coordinates (default : 0) to account
+        If necessary, apply a translation to streamline coordinates (default : 0) to account
         for differences between the reference system of the tracking algorithm and COMMIT.
         The value is specified in voxel units, eg 0.5 translates by half voxel.
 
@@ -94,13 +99,13 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
 
     blur_core_extent: float
         Extent of the core inside which the segments have equal contribution to the central one (default : 0.0).
-    
+
     blur_gauss_extent: float
         Extent of the gaussian damping at the border (default : 0.0).
 
     blur_gauss_min: float
         Minimum value of the Gaussian to consider when computing the sigma (default : 0.1).
-        
+
     blur_apply_to: array of bool
         For each input streamline, decide whether blur is applied or not to it (default : None, meaning apply to all).
 
@@ -135,7 +140,7 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     # check for invalid parameters in the blur
     if blur_core_extent < 0 :
         ERROR( 'The extent of the core must be non-negative' )
-    
+
     if blur_gauss_extent < 0 :
         ERROR( 'The extent of the blur must be non-negative' )
 
@@ -145,18 +150,18 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
 
     tic = time.time()
     LOG( '\n-> Creating the dictionary from tractogram:' )
-    
+
     LOG( '\n   * Configuration:' )
-    print( '\t- Segment position = %s' % ( 'COMPUTE INTERSECTIONS' if do_intersect else 'CENTROID' ) )
-    print( '\t- Fiber shift X    = %.3f (voxel-size units)' % fiber_shiftX )
-    print( '\t- Fiber shift Y    = %.3f (voxel-size units)' % fiber_shiftY )
-    print( '\t- Fiber shift Z    = %.3f (voxel-size units)' % fiber_shiftZ )
+    print( f'\t- Segment position = {"COMPUTE INTERSECTIONS" if do_intersect else "CENTROID"}' )
+    print( f'\t- Coordinates shift in X = {fiber_shiftX:.3f} (voxel-size units)' )
+    print( f'\t- Coordinates shift in Y = {fiber_shiftY:.3f} (voxel-size units)' )
+    print( f'\t- Coordinates shift in Z = {fiber_shiftZ:.3f} (voxel-size units)' )
     if min_seg_len >= 1e-3:
-        print( '\t- Min segment len  = %.3f mm' % min_seg_len )
+        print( f'\t- Min segment len  = {min_seg_len:.3f} mm' )
     else:
-        print( '\t- Min segment len  = %.2e mm' % min_seg_len )
-    print( '\t- Min fiber len    = %.2f mm' % min_fiber_len )
-    print( '\t- Max fiber len    = %.2f mm' % max_fiber_len )
+        print( f'\t- Min segment len  = {min_seg_len:.2e} mm' )
+    print( f'\t- Min streamline len    = {min_fiber_len:.2f} mm' )
+    print( f'\t- Max streamline len    = {max_fiber_len:.2f} mm' )
 
     # check blur params
     cdef :
@@ -166,8 +171,6 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
         bool [:] blurApplyTo
         int nReplicas
         float blur_sigma
-        float [:] ArrayInvM
-        float* ptrArrayInvM
 
     if (blur_gauss_extent==0 and blur_core_extent==0) or (blur_spacing==0) :
         nReplicas = 1
@@ -196,13 +199,13 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
                     blurWeights[i] = np.exp( -(blurRho[i] - blur_core_extent)**2 / (2.0*blur_sigma**2) )
 
     if nReplicas == 1 :
-        print( '\t- Do not blur fibers' )
+        print( '\t- Do not blur streamlines' )
     else :
-        print( '\t- Blur fibers:' )
-        print( '\t\t- core extent  = %.3f' % blur_core_extent )
-        print( '\t\t- gauss extent = %.3f (sigma = %.3f)' % (blur_gauss_extent, blur_sigma) )
-        print( '\t\t- grid spacing = %.3f' % blur_spacing )
-        print( '\t\t- weights = [ %.3f ... %.3f ]' % (np.min(blurWeights), np.max(blurWeights)) )
+        print( '\t- Blur streamlines:' )
+        print( f'\t\t- core extent  = {blur_core_extent:.3f}' )
+        print( f'\t\t- gauss extent = {blur_gauss_extent:.3f} (sigma = {blur_sigma:.3f})' )
+        print( f'\t\t- grid spacing = {blur_spacing:.3f}' )
+        print( f'\t\t- weights = [ {np.min(blurWeights):.3f} ... {np.max(blurWeights):.3f} ]' )
 
     if min_seg_len < 0 :
         ERROR( '"min_seg_len" must be >= 0' )
@@ -233,7 +236,7 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
         WARNING('"gen_trk" parameter is deprecated')
 
     # create output path
-    print( '\t- Output written to "%s"' % path_out )
+    print( f'\t- Output written to "{path_out}"' )
     if not exists( path_out ):
         makedirs( path_out )
 
@@ -244,16 +247,17 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
 
     # Streamlines from tractogram
     print( '\t- Tractogram' )
-    
+
     if not exists(filename_tractogram):
-        ERROR( 'Tractogram file not found: %s' % filename_tractogram )        
+        ERROR( f'Tractogram file not found: {filename_tractogram}' )
     extension = splitext(filename_tractogram)[1]
     if extension != ".trk" and extension != ".tck":
         ERROR( 'Invalid input file: only .trk and .tck are supported' )
-    
+
     hdr = nibabel.streamlines.load( filename_tractogram, lazy_load=True ).header
-            
+
     if extension == ".trk":
+        print ( f'\t\t- geometry taken from "{filename_tractogram}"' )
         Nx = hdr['dimensions'][0]
         Ny = hdr['dimensions'][1]
         Nz = hdr['dimensions'][2]
@@ -266,7 +270,7 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
         n_scalars = hdr['nb_scalars_per_point']
         n_properties = hdr['nb_properties_per_streamline']
 
-    if extension == ".tck":
+    else:
         if TCK_ref_image is None:
             if filename_peaks is not None:
                 TCK_ref_image = filename_peaks
@@ -274,49 +278,44 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
                 TCK_ref_image = filename_mask
             else:
                 ERROR( 'TCK files do not contain information about the geometry. Use "TCK_ref_image" for that' )
+        print ( f'\t\t- geometry taken from "{TCK_ref_image}"' )
 
-        print ('\t\t- geometry taken from "%s"' %TCK_ref_image)
-
-        nii_image = nibabel.load(TCK_ref_image)
-        nii_hdr = nii_image.header if nibabel.__version__ >= '2.0.0' else nii_image.get_header()
-        Nx = nii_image.shape[0]
-        Ny = nii_image.shape[1]
-        Nz = nii_image.shape[2]
-        Px = nii_hdr['pixdim'][1]
-        Py = nii_hdr['pixdim'][2]
-        Pz = nii_hdr['pixdim'][3]
-        data_offset = int(hdr['_offset_data'])  #set offset
-        n_count = int(hdr['count'])  #set number of fibers
+        niiREF = nibabel.load( TCK_ref_image )
+        niiREF_hdr = _get_header( niiREF )
+        Nx = niiREF.shape[0]
+        Ny = niiREF.shape[1]
+        Nz = niiREF.shape[2]
+        Px = niiREF_hdr['pixdim'][1]
+        Py = niiREF_hdr['pixdim'][2]
+        Pz = niiREF_hdr['pixdim'][3]
+        data_offset = int(hdr['_offset_data'])
+        n_count = int(hdr['count'])
         n_scalars = 0
         n_properties = 0
-        
-    print( '\t\t- %d x %d x %d' % ( Nx, Ny, Nz ) )
-    print( '\t\t- %.4f x %.4f x %.4f' % ( Px, Py, Pz ) )
-    print( '\t\t- %d fibers' % n_count )
+
+    print( f'\t\t- {Nx} x {Ny} x {Nz}' )
+    print( f'\t\t- {Px:.4f} x {Py:.4f} x {Pz:.4f}' )
+    print( f'\t\t- {n_count} streamlines' )
     if Nx >= 2**16 or Nz >= 2**16 or Nz >= 2**16 :
         ERROR( 'The max dim size is 2^16 voxels' )
-    
+
     # check copmpatibility between blurApplyTo and number of streamlines
     if blur_apply_to is None:
         blur_apply_to = np.repeat([True], n_count)
-    else : 
+    else :
         if blur_apply_to.size != n_count :
             ERROR( '"blur_apply_to" must have one value per streamline' )
-        print( '\t\t\t- %d blurred fibers' % sum(blur_apply_to) )
+        print( f'\t\t\t- {sum(blur_apply_to)} blurred streamlines' )
     blurApplyTo = blur_apply_to
-    
-    # get the affine matrix
-    if extension == ".tck":
-        scaleMat = np.diag(np.divide(1.0, [Px,Py,Pz]))
-        M = nii_hdr.get_best_affine()
 
-        # Affine matrix without scaling, i.e. diagonal is 1
-        M[:3, :3] = np.dot(scaleMat, M[:3, :3])
-        M = M.astype('<f4') # affine matrix in float value
-        invM = np.linalg.inv(M) # inverse affine matrix
-        #create a vector of inverse matrix M
-        ArrayInvM = np.ravel(invM)
-        ptrArrayInvM = &ArrayInvM[0]
+    # get toVOXMM matrix (remove voxel scaling from affine) in case of TCK
+    cdef float [:] toVOXMM
+    cdef float* ptrToVOXMM
+    if extension == ".tck":
+        M = _get_affine( niiREF ).copy()
+        M[:3, :3] = M[:3, :3].dot( np.diag([1./Px,1./Py,1./Pz]) )
+        toVOXMM = np.ravel(np.linalg.inv(M)).astype('<f4')
+        ptrToVOXMM = &toVOXMM[0]
 
     # white-matter mask
     cdef float* ptrMASK
@@ -324,9 +323,9 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     if filename_mask is not None :
         print( '\t- Filtering mask' )
         niiMASK = nibabel.load( filename_mask )
-        niiMASK_hdr = niiMASK.header if nibabel.__version__ >= '2.0.0' else niiMASK.get_header()
-        print( '\t\t- %d x %d x %d' % ( niiMASK.shape[0], niiMASK.shape[1], niiMASK.shape[2] ) )
-        print( '\t\t- %.4f x %.4f x %.4f' % ( niiMASK_hdr['pixdim'][1], niiMASK_hdr['pixdim'][2], niiMASK_hdr['pixdim'][3] ) )
+        niiMASK_hdr = _get_header( niiMASK )
+        print( f'\t\t- {niiMASK.shape[0]} x {niiMASK.shape[1]} x {niiMASK.shape[2]}' )
+        print( f'\t\t- {niiMASK_hdr["pixdim"][1]:.4f} x {niiMASK_hdr["pixdim"][2]:.4f} x {niiMASK_hdr["pixdim"][3]:.4f}' )
         if ( Nx!=niiMASK.shape[0] or Ny!=niiMASK.shape[1] or Nz!=niiMASK.shape[2] or
             abs(Px-niiMASK_hdr['pixdim'][1])>1e-3 or abs(Py-niiMASK_hdr['pixdim'][2])>1e-3 or abs(Pz-niiMASK_hdr['pixdim'][3])>1e-3 ) :
             WARNING( 'Dataset does not have the same geometry as the tractogram' )
@@ -342,17 +341,18 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
     cdef int Np
     cdef float [:, :, ::1] niiTDI_img = np.ascontiguousarray( np.zeros((Nx,Ny,Nz),dtype=np.float32) )
     cdef float* ptrTDI  = &niiTDI_img[0,0,0]
-    cdef double [:, ::1] affine
-    cdef double* ptrAFFINE
+    cdef double [:, ::1] peaksAffine
+    cdef double* ptrPeaksAffine
     if filename_peaks is not None :
         print( '\t- EC orientations' )
         niiPEAKS = nibabel.load( filename_peaks )
-        niiPEAKS_hdr = niiPEAKS.header if nibabel.__version__ >= '2.0.0' else niiPEAKS.get_header()
-        print( '\t\t- %d x %d x %d x %d' % ( niiPEAKS.shape[0], niiPEAKS.shape[1], niiPEAKS.shape[2], niiPEAKS.shape[3] ) )
-        print( '\t\t- %.4f x %.4f x %.4f' % ( niiPEAKS_hdr['pixdim'][1], niiPEAKS_hdr['pixdim'][2], niiPEAKS_hdr['pixdim'][3] ) )
-        print( '\t\t- ignoring peaks < %.2f * MaxPeak' % vf_THR )
-        print( '\t\t- %susing affine matrix' % ( "" if peaks_use_affine else "not " ) )
-        print( '\t\t- flipping axes : [ x=%s, y=%s, z=%s ]' % ( flip_peaks[0], flip_peaks[1], flip_peaks[2] ) )
+        niiPEAKS_hdr = _get_header( niiPEAKS )
+        print( f'\t\t- {niiPEAKS.shape[0]} x {niiPEAKS.shape[1]} x {niiPEAKS.shape[2]} x {niiPEAKS.shape[3]}' )
+        print( f'\t\t- {niiPEAKS_hdr["pixdim"][1]:.4f} x {niiPEAKS_hdr["pixdim"][2]:.4f} x {niiPEAKS_hdr["pixdim"][3]:.4f}' )
+
+        print( f'\t\t- ignoring peaks < {vf_THR:.2f} * MaxPeak' )
+        print( f'\t\t- {"" if peaks_use_affine else "not "}using affine matrix' )
+        print( f'\t\t- flipping axes : [ x={flip_peaks[0]}, y={flip_peaks[1]}, z={flip_peaks[2]} ]' )
         if ( Nx!=niiPEAKS.shape[0] or Ny!=niiPEAKS.shape[1] or Nz!=niiPEAKS.shape[2] or
             abs(Px-niiPEAKS_hdr['pixdim'][1])>1e-3 or abs(Py-niiPEAKS_hdr['pixdim'][2])>1e-3 or abs(Pz-niiPEAKS_hdr['pixdim'][3])>1e-3 ) :
             WARNING( "Dataset does not have the same geometry as the tractogram" )
@@ -366,15 +366,15 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
 
         # affine matrix to rotate gradien directions (if required)
         if peaks_use_affine :
-            affine = np.ascontiguousarray( niiPEAKS.affine[:3,:3].T )
+            peaksAffine = np.ascontiguousarray( niiPEAKS.affine[:3,:3].T )
         else :
-            affine = np.ascontiguousarray( np.eye(3) )
-        ptrAFFINE = &affine[0,0]
+            peaksAffine = np.ascontiguousarray( np.eye(3) )
+        ptrPeaksAffine = &peaksAffine[0,0]
     else :
         print( '\t- No dataset specified for EC compartments' )
         Np = 0
         ptrPEAKS = NULL
-        ptrAFFINE = NULL
+        ptrPeaksAffine = NULL
 
     # write dictionary information info file
     dictionary_info = {}
@@ -406,31 +406,33 @@ cpdef run( filename_tractogram=None, path_out=None, filename_peaks=None, filenam
         Nx, Ny, Nz, Px, Py, Pz, n_count, n_scalars, n_properties,
         fiber_shiftX, fiber_shiftY, fiber_shiftZ, min_seg_len, min_fiber_len, max_fiber_len,
         ptrPEAKS, Np, vf_THR, -1 if flip_peaks[0] else 1, -1 if flip_peaks[1] else 1, -1 if flip_peaks[2] else 1,
-        ptrMASK, ptrTDI, path_out, 1 if do_intersect else 0, ptrAFFINE,
-        nReplicas, &blurRho[0], &blurAngle[0], &blurWeights[0], &blurApplyTo[0], ptrArrayInvM, ndirs, ptrHashTable  );
+        ptrMASK, ptrTDI, path_out, 1 if do_intersect else 0, ptrPeaksAffine,
+        nReplicas, &blurRho[0], &blurAngle[0], &blurWeights[0], &blurApplyTo[0], ptrToVOXMM, ndirs, ptrHashTable  );
     if ret == 0 :
         WARNING( 'DICTIONARY not generated' )
         return None
 
     # save TDI and MASK maps
-    if filename_mask is not None :
-        affine = niiMASK.affine if nibabel.__version__ >= '2.0.0' else niiMASK.get_affine()
+    if TCK_ref_image is not None:
+        TDI_affine = _get_affine( niiREF )
+    elif filename_mask is not None :
+        TDI_affine = _get_affine( niiMASK )
     elif filename_peaks is not None :
-        affine = niiPEAKS.affine if nibabel.__version__ >= '2.0.0' else niiPEAKS.get_affine()
+        TDI_affine = _get_affine( niiPEAKS )
     else :
-        affine = np.diag( [Px, Py, Pz, 1] )
+        TDI_affine = np.diag( [Px, Py, Pz, 1] )
 
-    niiTDI = nibabel.Nifti1Image( niiTDI_img, affine )
-    nii_hdr = niiTDI.header if nibabel.__version__ >= '2.0.0' else niiTDI.get_header()
-    nii_hdr['descrip'] = 'Created with COMMIT %s'%get_distribution('dmri-commit').version
+    niiTDI = nibabel.Nifti1Image( niiTDI_img, TDI_affine )
+    niiTDI_hdr = _get_header( niiTDI )
+    niiTDI_hdr['descrip'] = f'Created with COMMIT {get_distribution("dmri-commit").version}'
     nibabel.save( niiTDI, join(path_out,'dictionary_tdi.nii.gz') )
 
     if filename_mask is not None :
-        niiMASK = nibabel.Nifti1Image( niiMASK_img, affine )
+        niiMASK = nibabel.Nifti1Image( niiMASK_img, TDI_affine )
     else :
-        niiMASK = nibabel.Nifti1Image( (np.asarray(niiTDI_img)>0).astype(np.float32), affine )
-    nii_hdr = niiMASK.header if nibabel.__version__ >= '2.0.0' else niiMASK.get_header()
-    nii_hdr['descrip'] = 'Created with COMMIT %s'%get_distribution('dmri-commit').version
+        niiMASK = nibabel.Nifti1Image( (np.asarray(niiTDI_img)>0).astype(np.float32), TDI_affine )
+    niiMASK_hdr = _get_header( niiMASK )
+    niiMASK_hdr['descrip'] = niiTDI_hdr['descrip']
     nibabel.save( niiMASK, join(path_out,'dictionary_mask.nii.gz') )
 
-    LOG( '\n   [ %.1f seconds ]' % ( time.time() - tic ) )
+    LOG( f'\n   [ {time.time() - tic:.1f} seconds ]' )